configApp/serializers/teach_serializer.py
from rest_framework import serializers

from configApp.models import Course


class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = ['id', 'title']
