requirements.txt
asgiref                   ~=     3.7.2
Django                   ~=      4.2.7
djangorestframework      ~=      3.14.0
djangorestframework-simplejwt ~=  5.3.0
drf-yasg          ~=             1.21.7
inflection       ~=              0.5.1
packaging        ~=              23.2
pip              ~=              23.0.1
PyJWT            ~=              2.8.0
pytz             ~=              2023.3.post1
PyYAML           ~=              6.0.1
setuptools       ~=              58.1.0
sqlparse          ~=             0.4.4
typing_extensions   ~=           4.8.0
uritemplate         ~=           4.1.1
django-cors-headers   ~=        4.3.1
django-filter         ~=          23.4
requests          ~=             2.31.0
python-dotenv     ~=             1.0.0
psycopg2-binary     ~=          2.9.9


